feat (extension): move data logic for authorized zkapp to route
